restful-pi/requirements.txt
aniso8601==7.0.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
click==7.0
flask-restplus==0.13.0
flask==1.1.1
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
jsonschema==3.0.2
markupsafe==1.1.1
pyrsistent==0.15.4
pytz==2019.2
requests==2.22.0
rpi.gpio==0.7.0
six==1.12.0
urllib3==1.25.3
werkzeug==0.15.5